embodied_agent_framework/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
pydantic>=2.0.0
python-dotenv>=0.19.0
pyyaml>=6.0
loguru>=0.6.0
asyncio-mqtt>=0.11.0

# Audio processing
pyaudio>=0.2.11
sounddevice>=0.4.4
wave>=0.0.2

# AI/ML dependencies
openai>=1.0.0
anthropic>=0.3.0
torch>=1.11.0
transformers>=4.20.0
sentence-transformers>=2.2.0

# Robot hardware (MyCobot)
pymycobot>=3.0.0

# Raspberry Pi GPIO (conditional)
RPi.GPIO>=0.7.1; platform_machine=="armv7l" or platform_machine=="aarch64"

# Web interface (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0